// src/features/dashboard/admin/types/meja.ts
export interface Meja {
  id: number;
  number: string;
  capacity: string;
  status: 'available' | 'occupied' | 'reserved' | 'maintenance';
  currentGuests: number;
  location: 'Indoor' | 'Outdoor';
  lastCleaned: string;
  reservedBy: string | null;
  reservedTime: string | null;
}

export interface CreateMejaInput {
  number: string;
  capacity: string;
  location: 'Indoor' | 'Outdoor';
}

// status default to 'available'
// reservedBy and reservedTime default to null
// id will be generated by the backend
//   lastCleaned will be a date string in ISO format and default date now in backend
//  currentGuests default to 0 in backend